% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horseshoe.R
\name{bjb_horseshoe}
\alias{bjb_horseshoe}
\title{horseshoe}
\usage{
bjb_horseshoe()
}
\arguments{
\item{W}{predictor matrix}

\item{z}{response variance}

\item{iteration}{Number of iterations}

\item{a, b}{rejection sampler parameter}

\item{s}{xi's hyperparameter}

\item{time_check}{Output run time}

\item{iteration_check}{Output iteration}

\item{beta}{the initial value of beta}

\item{xi}{the initial value of Global shrinkage parameter}

\item{Sigma}{the inial value of Sigma}

\item{w}{Sigma's hyperparameter}
}
\value{
beta posterior samples
}
\description{
Run linear model algorithm with horseshoe prior
}
